+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 All notable changes to `laravel-searchable` will be documented in this file
 
-## 1.7.1 - 2020-09-08
+## 1.9.0 - 2020-12-09
 
-- add support for PostgreSQL
+- add aspect search result limit (#82)
+
+## 1.8.0 - 2020-11-28
+
+- allow to use same query method (#81)
+
+## 1.7.1 - 2020-11-10
+
+- remove backticks in SearchAspect to support PostgreSQL (#85)
 
 ## 1.7.0 - 2020-09-08